fix: move navItem tree to extra file
